+++ CHANGELOG.md
@@ -42,6 +42,18 @@ Here are some of the features that notebooks bring:
 
   - After each cell is executed we issue a `retall` command to make sure that we are at the top-level and not stopped in a weird state
 
+## 3.2.5 October/November 2023
+
+Expanded sidebar for notebooks with buttons for:
+
+- Creating a new notebook
+
+- Formatting the IDL Code in your notebook cells (shortcut for VSCode native command)
+
+- Ability to open an example IDL Notebook
+
+There's also a command added that will reset your example notebooks to what we had originally. it is called "IDL: Reset IDL and ENVI Example Notebooks'.
+
 ## 3.2.4 October 2023
 
 Fixed an issue where improper versions of node.js were used for the language server startup. This caused failures in the language server starting which would prevent any of the feature goodness of the language server from being accessible.